+++ acts/src/package/event/manual.rs
@@ -10,7 +10,7 @@ use serde_json::json;
 use std::sync::Arc;
 
 #[derive(Debug, Clone, Serialize)]
-pub struct ManualEventPackage(serde_json::Value);
+pub struct ManualEventPackage(Option<Vars>);
 
 impl ActPackage for ManualEventPackage {
     fn meta() -> ActPackageMeta {
@@ -43,8 +43,8 @@ impl ActPackageFn for ManualEventPackage {
             )))?;
         let model: ModelInfo = rt.cache().store().models().find(&mid)?.into();
         let workflow = model.workflow()?;
-        let options = options.clone().with(consts::ACT_VALUE, self.0.clone());
-        let ret = rt.start(&workflow, &options)?;
+        let params = self.0.clone().unwrap_or(Vars::new());
+        let ret = rt.start(&workflow, &params)?;
 
         Ok(Some(Vars::new().with(consts::PROCESS_ID, ret.id())))
     }
@@ -55,7 +55,7 @@ impl<'de> serde::de::Deserialize<'de> for ManualEventPackage {
     where
         D: serde::Deserializer<'de>,
     {
-        let value = serde_json::Value::deserialize(deserializer)?;
+        let value = Option::<Vars>::deserialize(deserializer)?;
         Ok(Self(value))
     }
 }